- modify parser to enable sort command
