+++ src/main/java/seedu/doist/commons/core/Messages.java
@@ -4,11 +4,9 @@ package seedu.doist.commons.core;
  * Container for user visible messages.
  */
 public class Messages {
-
     public static final String MESSAGE_UNKNOWN_COMMAND = "Unknown command";
     public static final String MESSAGE_INVALID_COMMAND_FORMAT = "Invalid command format! \n%1$s";
     public static final String MESSAGE_INVALID_TASK_DISPLAYED_INDEX = "The task index provided is invalid";
     public static final String MESSAGE_TASKS_LISTED_OVERVIEW = "%1$d tasks listed!";
-    public static final String MESSAGE_TASKS_SORTED_OVERVIEW = "%1$d tasks sorted!";
-
+    public static final String MESSAGE_TASKS_SORTED_OVERVIEW = "Tasks sorted by %1$s!";
 }
