+++ src/main/java/seedu/doist/logic/commands/SortCommand.java
@@ -3,8 +3,6 @@ package seedu.doist.logic.commands;
 import java.util.ArrayList;
 import java.util.Arrays;
 
-import seedu.doist.commons.exceptions.IllegalValueException;
-
 
 /**
  * Sorts all persons in the to-do list by the specified parameter and shows it to the user.
@@ -22,7 +20,8 @@ public class SortCommand extends Command {
     public static ArrayList<String> commandWords =
             new ArrayList<>(Arrays.asList(DEFAULT_COMMAND_WORD, "sorted", "sort_by"));
 
-    public static final String MESSAGE_USAGE = DEFAULT_COMMAND_WORD + ":\n" + "Sorts previously listed tasks." + "\n"
+    public static final String MESSAGE_USAGE = DEFAULT_COMMAND_WORD
+            + ":\n" + "Sorts previously listed tasks." + "\n"
             + "Can sort by priority for now. \n\t"
             + "Parameters: SORTTYPE " + "\n\t"
             + "Example: " + DEFAULT_COMMAND_WORD
@@ -32,23 +31,8 @@ public class SortCommand extends Command {
             "You can only " + DEFAULT_COMMAND_WORD + "\n"
             + SortType.PRIORITY.toString();
 
-    public SortCommand(String argument) throws IllegalValueException {
-        // Remove trailing whitespace, spaces and change to upper case
-        String processedArgument = argument.trim();
-        processedArgument = processedArgument.replaceAll(" ", "");
-        processedArgument = processedArgument.toUpperCase();
-
-        if (!isValidSortArgument(processedArgument)) {
-            throw new IllegalValueException(MESSAGE_SORT_CONSTRAINTS);
-        }
-        this.sortType = SortType.valueOf(processedArgument);
-    }
-    /**
-     * Returns true if a given string is a valid sort argument
-     */
-    public static boolean isValidSortArgument(String argument) {
-        // Only priority for now
-        return argument.equals(SortType.PRIORITY.toString());
+    public SortCommand(SortType type) {
+        this.sortType = type;
     }
 
     @Override
